Refactored logger interfaces, extracted ObjectLoggerInterface and PropertyLoggerInterface
