+++ src/interfaces/ObjectLoggerInterface.ts
@@ -1,0 +1,7 @@
+import {LoggerInterface} from "./LoggerInterface";
+
+export interface ObjectLoggerInterface extends LoggerInterface {
+    timestampStart: number;
+    timestampEnd: number;
+    executionTime: () => number;
+}
