+++ src/interfaces/PropertyLoggerInterface.ts
@@ -1,0 +1,6 @@
+import {LoggerParams, LogOptions} from "./LoggerInterface";
+import {ObjectLoggerInterface} from "./ObjectLoggerInterface";
+
+export interface PropertyLoggerInterface extends ObjectLoggerInterface {
+    logPropertyChange: (propertyLogParams: LoggerParams, options?: LogOptions) => void;
+}
