+++ src/loggers/MethodLogger.ts
@@ -1,4 +1,5 @@
-import {LoggerParams, LoggerParamsType, LogOptions, MethodLoggerInterface} from "../interfaces/MethodLoggerInterface";
+import {LoggerParams, LoggerParamsType, LogOptions} from "../interfaces/LoggerInterface";
+import {MethodLoggerInterface} from "../interfaces/MethodLoggerInterface";
 import {ObjectLogger} from "./ObjectLogger";
 
 export class MethodLogger extends ObjectLogger implements MethodLoggerInterface {
